refactor givingsSection and apiTheGraph service
