+++ public/locales/en/translation.json
@@ -82,7 +82,8 @@
         "textLastCard": "Show your givings",
         "firstGivingButtonText": "Give now",
         "firstGivingText": "Make a giving and you’ll see it here",
-        "subtitleGivings": "Your givings",
+        "subtitleYourGivings": "Your givings",
+        "subtitleAllGivings": "All givings",
         "processingText": "Processing transaction...",
         "transactionSuccessText": "Transaction successful",
         "transactionFailedText": "Transaction failed",
