+++ public/locales/pt/translation.json
@@ -82,7 +82,8 @@
         "textLastCard": "Suas contribuições",
         "firstGivingButtonText": "Contribua agora",
         "firstGivingText": "Faça uma contribuição e veja ela aqui",
-        "subtitleGivings": "Suas contribuições",
+        "subtitleYourGivings": "Suas contribuições",
+        "subtitleAllGivings": "Todas as contribuições",
         "processingText": "Processando transação...",
         "transactionSuccessText": "Transação processada",
         "transactionFailedText": "Falha ao processar transação",
