+++ src/pages/promoters/FundPage/GivingsSection/index.test.tsx
@@ -1,12 +1,5 @@
-import { clickOn, renderComponent, waitForPromises } from "config/testUtils";
-import {
-  expectLogEventToHaveBeenCalledWith,
-  expectPageToNavigateTo,
-  expectTextToBeInTheDocument,
-} from "config/testUtils/expects";
-import promoterDonationFactory from "config/testUtils/factories/promoterDonationFactory";
-import { mockGraphqlRequest } from "config/testUtils/test-helper";
-import { PROMOTER_DONATIONS_QUERY_NAME } from "services/apiTheGraph/promoterDonationsApi";
+import { renderComponent } from "config/testUtils";
+import { expectTextToBeInTheDocument } from "config/testUtils/expects";
 import GivingsSection from ".";
 
 describe("GivingsSection", () => {
@@ -14,82 +7,6 @@ describe("GivingsSection", () => {
     renderComponent(<GivingsSection />);
 
     expectTextToBeInTheDocument("Your givings");
-  });
-
-  describe("when there is a wallet connected", () => {
-    describe("when the promoter has no givings", () => {
-      beforeEach(async () => {
-        renderComponent(<GivingsSection />, {
-          walletProviderValue: {
-            wallet: "0xFFFF",
-          },
-        });
-
-        await waitForPromises();
-      });
-
-      it("render make a giving card", () => {
-        expectTextToBeInTheDocument("Make a giving and you’ll see it here");
-      });
-
-      describe.skip("when the show your givings card is clicked", () => {
-        beforeEach(() => {
-          clickOn("Show your givings");
-        });
-
-        it("logs the fundShowGivingsListBtn_click event", () => {
-          expectLogEventToHaveBeenCalledWith("fundShowGivingsListBtn_click", {
-            from: "yourGivingsCarousel",
-          });
-        });
-
-        it("navigates to the you givings page", () => {
-          expectPageToNavigateTo("/promoters/show-givings");
-        });
-      });
-    });
-
-    describe("when the promoter has givings", () => {
-      beforeEach(async () => {
-        const fiftyCentInWei = "500000000000000000";
-        mockGraphqlRequest(PROMOTER_DONATIONS_QUERY_NAME, {
-          promoterDonations: [
-            promoterDonationFactory({
-              amountDonated: fiftyCentInWei,
-            }),
-          ],
-        });
-
-        renderComponent(<GivingsSection />, {
-          walletProviderValue: {
-            wallet: "0xFFFF",
-          },
-        });
-
-        await waitForPromises();
-      });
-
-      it("shows the user givings", async () => {
-        expectTextToBeInTheDocument("See transaction");
-        expectTextToBeInTheDocument("5/18/2022");
-        expectTextToBeInTheDocument("0.50");
-      });
-    });
-  });
-
-  describe("when there is no wallet connected", () => {
-    const connectWalletMock = jest.fn();
-
-    beforeEach(() => {
-      renderComponent(<GivingsSection />, {
-        walletProviderValue: {
-          connectWallet: connectWalletMock,
-        },
-      });
-    });
-
-    it("render make a giving card", () => {
-      expectTextToBeInTheDocument("Make a giving and you’ll see it here");
-    });
+    expectTextToBeInTheDocument("All givings");
   });
 });
