+++ src/services/apiTheGraph/promoterDonationsApi/index.tsx
@@ -1,38 +1,9 @@
-import { gql, ApolloQueryResult } from "@apollo/client";
+import { ApolloQueryResult } from "@apollo/client";
 import { client } from "..";
-
-export const PROMOTER_DONATIONS_QUERY_NAME = "PROMOTER_DONATIONS_QUERY";
-
-export const QUERY_PROMOTER_ID = gql`
-  query ${PROMOTER_DONATIONS_QUERY_NAME}($user: Bytes!, $first: Int) {
-    promoterDonations(
-      where: { user: $user }
-      orderBy: timestamp
-      orderDirection: desc
-      first: $first
-    ) {
-      id
-      user
-      amountDonated
-      timestamp
-    }
-  }
-`;
-
-export const QUERY_ALL_PROMOTERS = gql`
-  query ${PROMOTER_DONATIONS_QUERY_NAME}($first: Int) {
-    promoterDonations(
-      orderBy: timestamp
-      orderDirection: desc
-      first: $first
-    ) {
-      id
-      user
-      amountDonated
-      timestamp
-    }
-  }
-`;
+import {
+  QUERY_PROMOTER_DONATION_ID,
+  QUERY_ALL_PROMOTERS_DONATIONS,
+} from "../querys/promoterDonation";
 
 export const promoterDonationsApi = {
   fetchPromoterDonations: (
@@ -40,7 +11,7 @@ export const promoterDonationsApi = {
     first: number,
   ): Promise<ApolloQueryResult<any>> =>
     client.query({
-      query: QUERY_PROMOTER_ID,
+      query: QUERY_PROMOTER_DONATION_ID,
       variables: {
         user,
         first,
@@ -52,7 +23,7 @@ export const promoterDonationsApi = {
     first: number,
   ): Promise<ApolloQueryResult<any>> =>
     client.query({
-      query: QUERY_ALL_PROMOTERS,
+      query: QUERY_ALL_PROMOTERS_DONATIONS,
       variables: {
         first,
       },
